Place generated 404 at root of build folder
Previously 404.html was getting moved to 404/index.html which meant it woudn't be delivered to the user on sites where you can't control which file the server sends as a 404 (eg github pages). 

This change places it in the root of the build folder with a filename of 404.html, which matches most servers expected location.